% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wer.R
\name{wer}
\alias{wer}
\title{Calculating the Word Error Rate}
\usage{
wer(r, h)
}
\arguments{
\item{r}{The reference quanteda corpus}

\item{h}{The hypothesis quanteda corpus}
}
\description{
This function calculates the word error rate betwenn a hypothesis and a reference corpus.
}
\examples{
wer()
}
\keyword{error}
\keyword{rate}
\keyword{word}
