% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wersimtext.R
\name{wersimtext}
\alias{wersimtext}
\title{Run Text models on corpora that were modified using the WERSIM function}
\usage{
wersimtext(x, measured_wer, new_wer, deletions_sim = 0.13,
  insertions_sim = 0.22, substitutions_sim = 0.65, num_sims,
  preprocessing = c("punctuation", "numbers"), mincount_wersim = 0,
  method, groupingvar_sim, direction = c(1, 2))
}
\arguments{
\item{x}{A quanteda corpus to be modified}

\item{measured_wer}{The word error rate (or an estimate thereof) of corpus x}

\item{new_wer}{The word error rate for which the text model should be run. Generally, this is measured_wer plus some fixed incremental error such as 0.05}

\item{deletions_sim}{The share of word error that should be introduced through deletions}

\item{insertions_sim}{The share of word error that should be introduced through insertions}

\item{substitutions_sim}{The share of word error that should be introduced through substitutions}

\item{num_sims}{The number of simulations to be run}

\item{preprocessing}{The preprocessing that should be done with the corpus. Defaults to c("punctuation","numbers") for excluding numbers and punctuation, can take on "min_term", stemming", "stopwords_en", "stopwords_de".}

\item{mincount_wersim}{If "min_term" is part of "preprocessing", this parameters specifices the minimum number of times a word has to be in the corpus to be retained in the dfm}

\item{method}{The text model that should be run on the simulated corpus. Can either be "sentiment" or "wordfish"}

\item{groupingvar_sim}{The variable that groups the corpus}

\item{direction}{For Wordfish: The parameters that are forwarded to the dir command in Wordfish (fixing the direction of the space). Defaults to c(1,2).}
}
\value{
A data frame with the grouping variable in column 1 and simulated quantities (sentiment or Wordfish estimates) in subsequent columns.
}
\description{
This function calculates the word error rate between a hypothesis and a reference corpus.
}
\examples{
wersimtext()
}
\keyword{error}
\keyword{rate}
\keyword{transcriptions}
\keyword{wersim}
\keyword{word}
