% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wersim.R
\name{wersim}
\alias{wersim}
\title{Introducing additional Word Error into a corpus}
\usage{
wersim(x, target_wer = 0.05, deletions = 0.13, insertions = 0.22,
  substitutions = 0.65, groupingvar)
}
\arguments{
\item{x}{A quanteda corpus to be modified}

\item{target_wer}{The additional error that needs to be introduced into the corpus}

\item{deletions}{The share of word error that should be introduced through deletions}

\item{insertions}{The share of word error that should be introduced through insertions}

\item{substitutions}{The share of word error that should be introduced through substitutions}

\item{groupingvar}{The variable that groups the corpus}
}
\description{
This function introduces word error into a corpus
}
\examples{
wersim()
}
\keyword{error}
\keyword{rate}
\keyword{transcriptions}
\keyword{wersim}
\keyword{word}
